+++ src/util/board-state.ts
@@ -28,6 +28,7 @@ export enum StateChanges {
     // Board-wide kanban changes
     HISTORY_REVERSED,
     BOARD_LOADED,
+    NO_CHANGE_YET,
 }
 
 export type HistoryObject = {